update_loan no longer allows change of equip_ID

+++ loans/views.py
@@ -8,7 +8,7 @@ import loans.models
 from loans import models
 from .models import Equipment
 from .models import Loans
-from .forms import LoanForm
+from .forms import LoanForm, LoanForm1
  
 maxEquip = 0
 
@@ -48,31 +48,16 @@ def add_loan(request):
 
 def update_loan(request, id):
   loan = Loans.objects.get(id=id)
-  y = loan.equip_id
+  y=loan.equip_id
   if request.method == 'POST':
     form = LoanForm(request.POST, instance=loan)
     if form.is_valid():
-      #if form.cleaned_data['date_in'] is not None:
-      if loan.equip_id > 10 or loan.equip_id < 1:
-        return HttpResponse("That equipment id is invalid")
-      else:
+      if form.cleaned_data['equip_id'] != y:
+        return HttpResponse('<h2>You may not change equipment number.  Enter a return date, submit, and create a new loan transaction</h2>')
+      if form.cleaned_data['date_in'] is not None:
         x = Equipment.objects.get(id=form.cleaned_data['equip_id'])
-      if y == loan.equip_id:
-        #loan.save()
-        if loan.date_in:
-          x.checkedout = False
-        else:
-          x.checkedout = True
-        #x.save()
-      elif x.checkedout:
-        return HttpResponse("That equipment is already checked out")
-      # if equipment is available, save the form data to model
-      else:
-        if loan.date_in:
-          x.checkedout = False
-        else:
-          x.checkedout = True
-      x.save()
+        x.checkedout = False
+        x.save()
       form.save()
       return HttpResponseRedirect(reverse('home'))
   else:
@@ -123,6 +108,8 @@ def del_equip(request, id):
 
 def del_client(request, id):
   clients = Loans.objects.get(id=id)
+  if clients.date_in is None:
+    return HttpResponse('<h2>Delete record not authorized until equipment is checked in</h2')
   clients.delete()
   return HttpResponseRedirect(reverse('home'))
